+++ src/OpenTelemetry.Resources.Azure/CHANGELOG.md
@@ -13,6 +13,10 @@
   * `azure_vm` to `azure.vm`
   ([#5143](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5143))
 
+* Log an event when a resource detector does not detect any attributes because
+  the environment variable that identifies the platform is not set.
+  ([#5169](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5169))
+
 ## 1.18.0-beta.1
 
 Released 2026-Aug-21